third_party/blink/web_tests/svg/animations/animate-elem-03-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS parseFontSizeFromString() is 40
PASS colorComponents[1] is 0
PASS colorComponents[2] is 0
PASS colorComponents[3] is 255
PASS parseFontSizeFromString() is 60
PASS colorComponents[1] is 0
PASS colorComponents[2] is 59
PASS colorComponents[3] is 128
PASS parseFontSizeFromString() is 80
PASS colorComponents[1] is 0
PASS colorComponents[2] is 119
PASS colorComponents[3] is 0
PASS parseFontSizeFromString() is 80
PASS colorComponents[1] is 0
PASS colorComponents[2] is 119
PASS colorComponents[3] is 0
PASS successfullyParsed is true

TEST COMPLETE
